fix: Resolve circular import by renaming utils.py to table_utils.py
