+++ core/utils/__init__.py
@@ -2,7 +2,7 @@
 
 from .s3 import S3Utils
 
-# Import table utilities from the main utils module
-from ..utils import sanitize_table_name, create_table_name
+# Import table utilities from the table_utils module
+from ..table_utils import sanitize_table_name, create_table_name
 
 __all__ = ['S3Utils', 'sanitize_table_name', 'create_table_name']
